Address PR comments
* Move `ScopedDict` to `reframe/utility/__init__.py`.

* Add unit test for deleted parent and retaining nested scope.
